==> 01 - Introduction.ipynb <==
# FinancialAnalystAgent - An Agentic System to Investigate the use of LLMs in Financial Analysis

This project consists of multiple Jupyter Notebooks and python files that have been used to research three questions related to the use of LLMs to replicate the workflows of an equity financial analyst.

The three questions are:

1. Analysis of single LLM for stock price prediction and earnings estimates in line with Kim et al (2024). This will be compared to a baseline of consensus estimates from human analysts. We shall also discuss a comparison of open source and closed source models.
2. Analysis of fine-tuned single LLM to improve the performance of an open-sourced model for earnings estimates and come close to the ability of the model to closed-source.
3. Multi-agentic, multi-modal approach to improve the performance by including news and additional industry information in the model and a combination of open-source and closed source models.

This project requires additional modules to be installed. The cells below should be run depending on whether the project is being run for the Agentic system or run for the Analysis.


### Agentic Modules:

!pip -q install langgraph langchain_aws langchain-huggingface langchain

### Analysis Modules:

# Install signal_lab if it isn't already loaded
%package install bloomberg.bquant.signal_lab=1.5.2
