==> 01A - Financial Datasets.ipynb <==
# Datasets

This notebook is the first step in running the Financial Analyst workflow and is needed to request all of the data in order to run the strategy construction processes later. The data retrieval component is time intensive and so we pre-request the datasets, store them in Bloomberg Lab S3 Storage for lower latency access during the construction process.

This notebook request the following datasets:

- Financial Statement Point in time datasets from Bloomberg Query Language (BQL)
- Datasets needed to run the Portfolio Construction process to view the performance of the strategies from the LLMs.

For the analysis in this report, we shall be requesting financial statement datasets for the Dow Jones index from 2020-01-01 to 2025-03-31. The helper file in the 'requesters' folder makes the requests and downloads the datasets into a format ready to use by the Strategy Construction module.

The code base for this module is in the requesters folder. 

import bql
import os
import importlib
import json

import pandas as pd

# All of the custom requesters to help with data retrieval
import requesters.data_request_helper as helper
from requesters.company_data import SecurityData
from requesters.data_request_helper import FinancialDataRequester

from utils.s3_helper import S3Helper
from prompts import SYSTEM_PROMPTS

## Quarterly Data for the Dow Jones
Configure the variables for the index - INDU Index is the Bloomberg identifier for the Dow Jones Index, a filename to store the data in Bloomberg Lab S3 Storage, a reporting period which will be the quarterly reporting datasets, a start date and a set of rebalance dates.

The rebalance dates are used to check the index on a quarterly basis for changes.

# Index to use for point in time firms
index = 'INDU Index'
filename = 'data_quarterly.json'
reporting_period = 'Q'
start_date = '2020-01-01'

# rebalance dates for the index
rebalance_dates = ['2024-12-31',
        '2024-09-30',
        '2024-06-30',
        '2024-03-31',
        '2023-12-31',
        '2023-09-30',
        '2023-06-30',
        '2023-03-31',
        '2022-12-31',
        '2022-09-30',
        '2022-06-30',
        '2022-03-31',
        '2021-12-31',
        '2021-09-30',
        '2021-06-30',
        '2021-03-31',
        '2020-12-31',
        '2020-09-30',
        '2020-06-30',
        '2020-03-31',
        '2019-12-31',
        '2019-09-30',
        '2019-06-30',
        '2019-03-31']

# Set up the data_helper object using the Financial Data Requester which will format the datasets into the correct format for us to use in the strategy construction module.
data_helper = helper.FinancialDataRequester(index_id=index,
                                            dataset_name='quarterly_pit_indu_blended',
                                            rebalance_dates=rebalance_dates,
                                            reporting_frequency=reporting_period,
                                            start_date=start_date)

df_rebalance_dates = data_helper.get_rebalance_dates()

df_rebalance_dates

all_data = data_helper.create_financial_dataset()

#### Save to Bloomberg Lab S3 Storage

# Write the data to local ephemeral storage
local_file = '/tmp/dow_quarterly_ltm_v3.json'
with open(local_file, 'w') as f:
    json.dump(all_data, f)

# Create S3 Helper object
s3_helper = S3Helper('tmp/fs')

# Upload to Bloomberg Lab S3 Storage
s3_helper.add_file(local_filename=local_file)

## Annual Data for the Dow Jones

# Index to use for point in time firms
index = 'INDU Index'
filename = 'data_annual_pit_dow.json'
reporting_period = 'A'

# rebalance dates for the index
rebalance_dates = ['2024-12-31',
        '2024-09-30',
        '2024-06-30',
        '2024-03-31',
        '2023-12-31',
        '2023-09-30',
        '2023-06-30',
        '2023-03-31',
        '2022-12-31',
        '2022-09-30',
        '2022-06-30',
        '2022-03-31',
        '2021-12-31',
        '2021-09-30',
        '2021-06-30',
        '2021-03-31',
        '2020-12-31',
        '2020-09-30',
        '2020-06-30',
        '2020-03-31',
        '2019-12-31',
        '2019-09-30',
        '2019-06-30',
        '2019-03-31',]

data_helper = helper.FinancialDataRequester(index_id=index,
                                            dataset_name='annual_pit_indu_blended',
                                            rebalance_dates=rebalance_dates,
                                            reporting_frequency=reporting_period)

all_data = data_helper.create_financial_dataset()

all_data['2020-04-24']['AXP UN Equity']['mt']

## S&P 500 Datasets for LLM Training

# select the index
training_index = 'SPX Index'
filename = 'data_quarterly_pit_spx_refresh_blended.json'
reporting_period = 'Q'
start_date = '2020-01-01'

# rebalance dates for the index
rebalance_dates = ['2024-12-31',
        '2024-09-30',
        '2024-06-30',
        '2024-03-31',
        '2023-12-31',
        '2023-09-30',
        '2023-06-30',
        '2023-03-31',
        '2022-12-31',
        '2022-09-30',
        '2022-06-30',
        '2022-03-31',
        '2021-12-31',
        '2021-09-30',
        '2021-06-30',
        '2021-03-31',
        '2020-12-31',
        '2020-09-30',
        '2020-06-30',
        '2020-03-31',
        '2019-12-31',
        '2019-09-30',
        '2019-06-30',
        '2019-03-31',]

data_helper = helper.FinancialDataRequester(index_id=index,
                                            dataset_name=filename,
                                            rebalance_dates=rebalance_dates,
                                            reporting_frequency=reporting_period,
                                            start_date=start_date)

training_data = data_helper.create_financial_dataset()

#### Save to S3

# Write the data to local ephemeral storage
local_file = f'/tmp/{filename}'
with open(local_file, 'w') as f:
    json.dump(all_data, f)

# Upload to Bloomberg Lab S3 Storage
s3_helper.add_file(local_filename=local_file)

### Example prompt
Below is an example of the Income Statement, Balance Sheet and Historical price data that has been generated from the SecurityData class. A more detailed breakdown of the Security Data class can be found in the Tests folder of this project.

security_data = SecurityData('tmp/fs','dow_quarterly_ltm_v3.json')

prompt = security_data.get_prompt('2020-01-08', 'WBA UW Equity', SYSTEM_PROMPTS['BASE']['prompt'])

prompt

## Request Data for Strategy Analytics

We also need datasets to properly evaluate the strategies created. We use Signal Lab libraries to help with requesting returns datasets, datasets for the benchmark and attribution data. This is in the event_study.py project.

from analytics.event_study import EventBacktest
from bloomberg.bquant.signal_lab.workflow.utils import get_sandbox_path

import importlib

data_pack_path= f"{get_sandbox_path()}/tmp/fs/test_datapack"
backtest = EventBacktest(start=start_date, 
                         end='2025-03-31' , 
                         universe_name=index, 
                         data_pack_path=data_pack_path, 
                         reload_data=True)
