==> 01C - News Datasets.ipynb <==
# News Article Data Request
This code is used to request Bloomberg news headlines data. In BQuant Enterprise, this is accessed in csv format over a period from 2008 to today. It is recommended to load the datasets using Apache Spark as they are very large datasets. 

This must be requested in the notebook so that the spark session can be correctly configured and shut down. In this notebook, we use the S3Helper to save the data into BQuant Enterprise storage so that it can be used in other parts of the agentic system. 

from pyspark.sql import SparkSession, Window
import pyspark.sql.functions as F
from pyspark.sql.types import TimestampType

from utils.s3_helper import S3Helper
from requesters.company_data import SecurityData
from IPython.display import display

import os
import boto3

### Set up and configure Spark
We set up Spark with 100 workers, each with 2 vCPUs and 8GB of memory. 


def get_spark_session(
    executors="10",
    executor_memory="8g",
    driver_memory="32g",
    executor_cores="2",
    driver_max_result_size="1024M",
    executor_memory_overhead="2g",
    task_cpus="1",
):
    """
    Function to set up and define the Spark session. 
    """

    spark = (
        SparkSession.builder.config("spark.driver.memory", driver_memory)
        .config("spark.driver.maxResultSize", driver_max_result_size)
        .config("spark.executor.memoryOverhead", executor_memory_overhead)
        .config("spark.executor.instances", executors)
        .config("spark.executor.memory", executor_memory)
        .config("spark.executor.cores", executor_cores)
        .config("spark.task.cpus", task_cpus)
        .config("spark.sql.execution.arrow.enabled", "true")
        .config("spark.shuffle.file.buffer", "1m")
        .config("spark.file.transferTo", "False")
        .config("spark.shuffle.unsafe.file.output.buffer", "1m")
        .config("spark.io.compression.lz4.blockSize", "512k")
        .config("spark.shuffle.service.index.cache.size", "1g")
        .config("spark.shuffle.registration.timeout", "120000ms")
        .config("spark.shuffle.registration.maxAttempts", "3")
        .config("spark.sql.windowExec.buffer.spill.threshold", "1000000")
        .config("spark.sql.windowExec.buffer.in.memory.threshold", "1000000")
        .getOrCreate()
    )

    display(spark)

    return spark

# Set up a spark session with 100 executors on 2 core machines.
spark = get_spark_session(executors="100", executor_memory="8g", executor_cores="2")

### Request News Headlines datasets 
Load all of the Bloomberg News headlines datasets into spark. The files are stored as daily snapshots. The workers will load each csv file and load into a Spark DataFrame for processing later. Loading all files in this case took 6min 11s to complete.

We are only going to request headlines data for this project. It is possible to collect the news bodies as well but this will take considerably longer to request and would be difficult for us to fit into memory with the LLMs. This is a future areas for research but out of scope for our project.

%%time
bucket_name = "bquant-data-textual-analytics-tier-1"
bucket = boto3.resource("s3").Bucket(bucket_name)
files = [file.key for file in bucket.objects.all()]

files_csv = [
    f"s3://{bucket_name}/{file}"
    for file in files
    if "EID80001" in file and "csv" in file
]

df = (
    spark.read.option("header", "true")
    .option("multiLine", "true")
    .option("escape", "")
    .csv(files_csv)
)

### Filter the news articles
We are going to manually filter for news headlines on the Filter based on the FIGI identifiers for the Dow Jones, articles after 1st January 2019 and Bloomberg News articles only. We will only use Bloomberg News headlines because of 3rd party news source limitations on use with LLMs. 

For future research, we could also filter for economic news and add a macro-economic element to our agentic model. We filter for English language headlines and only headlines longer than 25 characters (smaller headlines are likely to be highly abbreviated and add little value to the trend prediction process)

sec_helper = SecurityData('tmp/fs','dow_quarterly_ltm_v3.json')
index_members = sec_helper.get_unique_securities(is_figi=True)
index_members[:10]

%%time
from functools import reduce

start_date = '2019-01-01'
# Filter for just BBG news or include all news articles in the analysis.
wire_filter = (F.col("WireName") == "BN") | (F.col("WireName") == "BFW")

filters = (
    wire_filter
    & (F.col("LanguageString") == "ENGLISH")
    & (F.length(F.col("Headline")) > 25)
    & (F.col("TimeOfArrival") >= start_date)
    & (F.col("Assigned_ID_BB_GLOBAL").isin(index_members))
)

df = df.withColumn("TimeOfArrival", F.col("TimeOfArrival").cast(TimestampType()))
df1 = df.filter(filters)

df1 = df1.cache()
df1.count()

### Drop Duplicates

window = Window.partitionBy("SUID").orderBy(F.col("TimeOfArrival").asc())

df2 = (
    df1.withColumn("row", F.row_number().over(window))
    .filter(F.col("row") == 1)
    .drop("row")
)

df2 = df2.withColumn("day", F.to_date(F.col("TimeOfArrival")))
window = Window.partitionBy("day", "Headline").orderBy(F.col("TimeOfArrival").asc())
df2 = (
    df2.withColumn("row", F.row_number().over(window))
    .filter(F.col("row") == 1)
    .drop("row", "day")
)

df2 = df2.cache()
df2.count()

### Sort the Headlines
Reduce the dataframe to News headline ID, headline text, the date of the headline and the company associated with the headline. 

def sort_headlines(df2):
    pdf = (
        df2.select(
            "SUID", "Headline", "TimeOfArrival", "Assigned_ID_BB_GLOBAL"
        )
        .toPandas()
        .sort_values(by="TimeOfArrival")
        .reset_index(drop=True)
        .copy()

    )
    pdf["Headline"] = pdf["Headline"].str.lower()

    return pdf

headlines = sort_headlines(df2)


headlines['Headline'][0]

### Output Headlines to S3

s3_helper = S3Helper('tmp/fs')

os.mkdir('/tmp/headlines')
headlines.to_parquet('/tmp/headlines/dow_headlines.parquet')

s3_helper.add_file(local_filename='/tmp/headlines/dow_headlines.parquet')

### Stop Spark

spark.stop()

import pandas as pd
import plotly.express as px

headlines

pd.set_option('display.max_colwidth', 350)
