==> 02 - Open Source Inference.ipynb <==
# Run Open Source Inference 

This notebook runs inference on the open source models using the multi-gpu process implemented in the modelinference file. This file requires the llm-base environment to run correctly. The results of the backtest are stored in the Data folder under each model folder. 

We test with a handful of open source models:

- Llama 3.2 3B Instruct
- Qwen 2.5 7B Instruct
- Deepseek R1 Qwen 14B 
- Qwen 2.5 3B Instruct

We also test with a fine-tuned model, the Qwen 2.5 3B model. We test this with the earnings prompt and also with a new EPS estimate prompt before fine-tuning to provide a comparison. 

#!pip install triton

#import model_inference
import prompts
import importlib
import datetime
import torch
import json
from huggingface_hub import login
import constructors.huggingface_strategy as hs

import models.model_helper as mh

from constructors.huggingface_strategy import HuggingfaceRun
from transformers import BitsAndBytesConfig
from accelerate import Accelerator, notebook_launcher

importlib.reload(prompts)
importlib.reload(mh)
importlib.reload(hs)

### Set up the environment
Log into Huggingface and check the number of GPUs available, pytorch version currently loaded

# Log into Huggingface with login credentials
with open('pass.txt') as p:
    hf_login = p.read()
    
hf_login = hf_login[hf_login.find('=')+1:hf_login.find('\n')]
login(hf_login, add_to_git_credential=False)

print(f'Torch version: {torch.__version__}')
#print(f'Device Count: {torch.cuda.device_count()}')
import accelerate
accelerate.__version__

### Run 1: Llama 3.2 Earning analysis - Base
The first model we test is the smallest one - Llama 3.2 3B model. As a Huggingface model we use the HuggingfaceRun object to create the strategy and produce the list of trades from the run. We must launch in notebook_launcher to use multi-GPUs. 

Results File: Results/Earnings/results - llama - base - 2025-03-30 896587.json

# Create the run config
run_config = {
    'model_s3_loc': 'llama',
    'model_reload': False,
    'model_quant': None,
    'multi-gpu':True,
    'dataset': 'data_quarterly_pit_indu_blended',
    'data_location': 'data_quarterly_pit_indu_refresh_blended.json'

}

ir = hs.HuggingfaceRun(run_name='Llama 3B Earnings',
                       model_id='meta-llama/Llama-3.2-3B-Instruct',
                       dataset_id='data_quarterly_pit_indu_refresh_blended.json',system_prompt=prompts.SYSTEM_PROMPTS['BASE_EARN'],run_config=run_config)

# Create the prompts and save to the Data folder
prompt_set = ir.create_all_prompts(force_refresh=True, is_save_prompts=True)

# Run the multi-gpu model with notebook_launcher
notebook_launcher(ir.run_multi_gpu, num_processes=torch.cuda.device_count())

### Run 2: Llama 3.2 Earning analysis - Chain of Thought

Run the Llama 3.2 model on Chain of Thought prompts. Results in the following folder:

Results/Earnings/results - llama - cot - 2025-03-30 332729.json


run_config = {
    'model_s3_loc': 'llama',
    'model_reload': False,
    'model_quant': None,
    'multi-gpu':True,
    'data_location': 'data_quarterly_pit_indu_refresh_blended.json'
}

ir = HuggingfaceRun(run_name='Llama 3B Earnings COT',
                       model_id='meta-llama/Llama-3.2-3B-Instruct',
                       dataset_id='data_quarterly_pit_indu_refresh_blended.json',
                       system_prompt=prompts.SYSTEM_PROMPTS['COT_EARN'],run_config=run_config)

# Create the prompts and save to the Data folder
prompt_set = ir.create_all_prompts(force_refresh=True, is_save_prompts=True)

# Run the multi-gpu model with notebook_launcher
notebook_launcher(ir.run, num_processes=torch.cuda.device_count())

### Run 3: DeepSeek R1 Qwen 7B Base

Results in the following file:

Results/Earnings/results - Deepseek - BASE C -2025-04-01 20:43:33.936328.json

run_config = {
    'model_s3_loc': 'deepseek7B',
    'model_reload': False,
    'model_quant': None,
    'multi-gpu':True
}

ir = HuggingfaceRun(run_name='DeepSeek 7B Earnings Base',
                       model_id='deepseek-ai/DeepSeek-R1-Distill-Qwen-7B',
                       dataset_id='data_quarterly_pit_indu_refresh_blended.json',
                       system_prompt=prompts.SYSTEM_PROMPTS['BASE_EARN'],run_config=run_config)

# Create the prompts and save to the Data folder
prompt_set = ir.create_all_prompts(force_refresh=True, is_save_prompts=True)

# Run the multi-gpu model with notebook_launcher
notebook_launcher(ir.run_multi_gpu, num_processes=torch.cuda.device_count())

### Run 4: DeepSeek 7B COT

Results in the following file:

Results/Earnings/results - Deepseek - COT C -2025-04-01 21:13:35.944412.json

run_config = {
    'model_s3_loc': 'deepseek7B',
    'model_reload': False,
    'model_quant': None,
    'multi-gpu':True
}

ir = HuggingfaceRun(run_name='DeepSeek 7B Earnings COT',
                       model_id='deepseek-ai/DeepSeek-R1-Distill-Qwen-7B',
                       dataset_id='data_quarterly_pit_indu_refresh_blended.json',
                       system_prompt=prompts.SYSTEM_PROMPTS['COT_EARN'],run_config=run_config)

# Create the prompts and save to the Data folder
prompt_set = ir.create_all_prompts(force_refresh=True, is_save_prompts=True)

notebook_launcher(ir.run_multi_gpu, num_processes=torch.cuda.device_count())

### Qwen 2.5 - 3B Instruct
For this model, we apply quantization to reduce the size of the model. This will be used for fine-tuning and so memory footprint needs to be smaller to permit training on a single GPU. We use the Bits and Bytes library from Huggingface to convert to 4bit.

run_config = {
    'model_s3_loc': 'qwen3b',
    'model_reload': False,
    'model_quant': None,
    'multi-gpu':True,
}

ir = HuggingfaceRun(run_name='Qwen 3B Earnings BASE',
                       model_id='Qwen/Qwen2.5-3B-Instruct',
                       dataset_id='data_quarterly_pit_indu_refresh_blended.json',
                       system_prompt=prompts.SYSTEM_PROMPTS['BASE_EARN'],run_config=run_config)

# Create the prompts and save to the Data folder
prompt_set = ir.create_all_prompts(force_refresh=True, is_save_prompts=True)

notebook_launcher(ir.run_multi_gpu, num_processes=torch.cuda.device_count())

run_config = {
    'model_s3_loc': 'qwen3b',
    'model_reload': False,
    'model_quant': None,
    'multi-gpu':True
}

ir = HuggingfaceRun(run_name='Qwen 3B Earnings COT',
                       model_id='Qwen/Qwen2.5-3B-Instruct',
                       dataset_id='data_quarterly_pit_indu_refresh_blended.json',
                       system_prompt=prompts.SYSTEM_PROMPTS['COT_EARN'],run_config=run_config)

# Create the prompts and save to the Data folder
prompt_set = ir.create_all_prompts(force_refresh=True, is_save_prompts=True)

notebook_launcher(ir.run_multi_gpu, num_processes=torch.cuda.device_count())

## Run 5 - Fine tuned model - EPS only

run_config = {
    'model_hf_id': 'Qwen/Qwen2.5-3B-Instruct',
    'model_s3_loc': 'qwen3b',
    'model_reload': False,
    'model_quant': None,
    'system_prompt': prompts.SYSTEM_PROMPTS['BASE_FINE_TUNED'],
    'multi-gpu':False,
    'dataset': 'data_quarterly_pit_indu_blended_base',
    'data_location': 'data_quarterly_pit_indu_refresh_blended.json',
    'fine_tuned_dir': 'fine_tuned2'
}

ftm = mft.FineTunedInference("EPS only FT Run", run_config)

prompt_set = ftm.create_all_prompts(force_refresh=True, is_save_prompts=True)
prompt_set_appended = ftm.reformat_prompts(prompt_set, "The next period EPS is ")

output_eps = ftm.run_finetuned_backtest(prompt_set_appended)

# save the output
with open(f"Results/Earnings/results - Qwen3B Finetuned - EPS only.json", 'w') as f:
    json.dump(output_eps, f)

ir = HuggingfaceRun(run_name='Qwen Earnings',
                       model_id='Qwen/Qwen2.5-3B-Instruct',
                       dataset_id='data_quarterly_pit_indu_refresh_blended.json',
                       system_prompt=prompts.SYSTEM_PROMPTS['BASE_FINE_TUNED'],run_config=run_config)

prompt_set = ir.create_all_prompts(force_refresh=True, is_save_prompts=True)

notebook_launcher(ir.run_multi_gpu, num_processes=torch.cuda.device_count())

output = ir.cached_results

## Stock Recommendation Runs


run_config = {
    'model_hf_id': 'deepseek-ai/DeepSeek-R1-Distill-Qwen-14B',
    'model_s3_loc': 'deepseek14Q',
    'model_reload': False,
    'model_quant': None,
    'system_prompt': prompts.SYSTEM_PROMPTS['CoT'],
    'multi-gpu':True,
    'dataset': 'data_annual_pit_spx',
    'data_location': 'data_annual_pit_spx.json'
}

run_config = {
    'model_hf_id': 'Qwen/Qwen2.5-7B-Instruct',
    'model_s3_loc': 'qwen',
    'model_reload': False,
    'model_quant': None,
    'system_prompt': prompts.SYSTEM_PROMPTS['BASE'],
    'multi-gpu':True,
    'dataset': 'data_quarterly_pit_indu',
    'data_location': 'data_quarterly_pit_indu.json'
}

# RUN 1
run_config = {
    'model_hf_id': 'meta-llama/Llama-3.2-3B-Instruct',
    'model_s3_loc': 'llama',
    'model_reload': False,
    'model_quant': None,
    'system_prompt': prompts.SYSTEM_PROMPTS['BASE'],
    'multi-gpu':True,
    'dataset': 'data_quarterly_pit_indu',
    'data_location': 'data_quarterly_pit_indu.json'
}

# RUN 2
run_config = {
    'model_hf_id': 'meta-llama/Llama-3.2-3B-Instruct',
    'model_s3_loc': 'llama',
    'model_reload': False,
    'model_quant': None,
    'system_prompt': prompts.SYSTEM_PROMPTS['CoT'],
    'multi-gpu':True,
    'dataset': 'data_quarterly_pit_indu',
    'data_location': 'data_quarterly_pit_indu.json'
}

# RUN 3
run_config = {
    'model_hf_id': 'deepseek-ai/DeepSeek-R1-Distill-Qwen-14B',
    'model_s3_loc': 'deepseek14Q',
    'model_reload': False,
    'model_quant': None,
    'system_prompt': prompts.SYSTEM_PROMPTS['BASE'],
    'multi-gpu':True,
    'dataset': 'data_quarterly_pit_indu',
    'data_location': 'data_quarterly_pit_indu.json'
}

# RUN 4
run_config = {
    'model_hf_id': 'deepseek-ai/DeepSeek-R1-Distill-Qwen-14B',
    'model_s3_loc': 'deepseek14Q',
    'model_reload': False,
    'model_quant': None,
    'system_prompt': prompts.SYSTEM_PROMPTS['CoT'],
    'multi-gpu':True,
    'dataset': 'data_quarterly_pit_indu',
    'data_location': 'data_quarterly_pit_indu.json'
}


# Run 6 - failed
run_config = {
    'model_hf_id': 'Qwen/Qwen2.5-7B-Instruct',
    'model_s3_loc': 'qwen',
    'model_reload': False,
    'model_quant': None,
    'system_prompt': prompts.SYSTEM_PROMPTS['CoTDetailed'],
    'multi-gpu':True,
    'dataset': 'data_quarterly_pit_indu',
    'data_location': 'data_quarterly_pit_indu.json'
}

run_config = {
    'model_hf_id': 'meta-llama/Llama-3.2-3B-Instruct',
    'model_s3_loc': 'llama',
    'model_reload': False,
    'model_quant': None,
    'system_prompt': prompts.SYSTEM_PROMPTS['BASE_EARN'],
    'multi-gpu':True,
    'dataset': 'data_quarterly_pit_indu_refresh_v2',
    'data_location': 'data_quarterly_pit_indu_refresh_v2.json'
}




run_name = f"{run_config['model_s3_loc']}_{run_config['dataset']}"

ir = model_inference.InferenceRun(run_name, run_config)

# Create the prompts and save to the Data folder
prompts = ir.create_all_prompts(True)

# Run the multi-gpu model with notebook_launcher
notebook_launcher(ir.run_multi_gpu, num_processes=torch.cuda.device_count())

ir = model_inference.InferenceRun(run_name, run_config)

p1 = ir.create_all_prompts(True)

len(p1)

model = ir.load_model_from_storage(ir.model_s3_loc)

from transformers import AutoTokenizer

tokenizer = AutoTokenizer.from_pretrained(ir.model_hf_id)

output_result = ir.run_model(p1[0]['prompt'],tokenizer,model)

output_result = {"date": "2020-02-06", "security": "MMM UN Equity", "response": "Here is the JSON response:\n\n```json\n{\n  \"decision\": \"BUY\",\n  \"confidence score\": 80,\n  \"reason\": \"Gross profit and EPS have increased over time, indicating a strong financial performance\"\n}\n```\n\nI have computed the following financial ratios:\n\n1. Gross Margin: \n   - 2020: 3.786000e+09 / 8.111000e+09 = 0.466\n   - 2019: 3.803000e+09 / 7.991000e+09 = 0.477\n   - 2018: 3.858000e+09 / 8.171000e+09 = 0.471\n   - 2017: 3.553000e+09 / 7.863000e+09 = 0.453\n   - 2016: 3.885000e+09 / 7.945000e+09 = 0.492\n\nThe gross margin has been increasing over time, indicating a strong financial performance.\n\n2. EPS:\n   - 2020: 9.690000e+08 / 1.012600e+10 = 0.953\n   - 2019: 1.583000e+09 / 1.076400e+10 = 1.465\n   - 2018: 1.127000e+09 / 1.014200e+10 = 1.117\n   - 2017: 1.347000e+09 / 9.848000e+09 = 0.137\n   - 2016: 1.543000e+09 / 9.848000e+09 = 0.156\n\nThe EPS has been increasing over time, indicating a strong financial performance.\n\n3. Current Ratio:\n   - 2020: 2.441000e+09 / 9.222000e+09 = 0.264\n   - 2019: 1.588000e+09 / 7.821000e+09 = 0.201\n   - 2018: 1.131000e+09 / 7.265000e+09 = 0.157\n   - 2017: 8.930000e+08 / 7.244000e+09 = 0.123\n   - 2016: 8.910000e+08 / 5.020000e+09 = 0.177\n\nThe current ratio has been increasing over time, indicating a strong financial performance.\n\nBased on these financial ratios, the company has been performing well financially and has a strong track record of increasing gross profit and EPS over time. Therefore, I recommend a BUY decision with a confidence score of 80."}

output_result['response']

import json

def format_json(llm_output):
    # remove all the broken lines
    form = llm_output.replace('\n','')
    # Find the start and end of the JSON input
    #try:
    soj = form.find('```json')
    eoj = form.find('}```')

    if eoj == -1:
        eoj = len(llm_output)
        llm_output = llm_output + '}```'
    # Pull out the additional context
    additional = form[:soj]
    additional += form[eoj + 4:]
    json_obj = json.loads(form[soj + 7:eoj + 1])
    json_obj['AdditionalContext'] = additional
    return json_obj
    #except:
    #    return llm_output

format_json(output_result['response'])

ir.save_run({'test':'1234'})

# Old Code
# run_name = f"{run_config['model_s3_loc']}_{run_config['dataset']}_finetuned"
# ir = model_inference.InferenceRun(run_name, run_config)

# # Create the prompts and save to the Data folder
# prompt_set = ir.create_all_prompts(force_refresh=True, is_save_prompts=True)
# for prompt in prompt_set:
#     prompt['prompt'] += 'The next period EPS is '
#     #prompt['prompt'] += "\nAnswer in JSON format with the next period EPS, the direction, the magnitude and a confidence."

# from peft import PeftModel
# from transformers import AutoTokenizer
# import json
# from tqdm import tqdm

# outputs = []
# def run_finetuned_backtest(prompts, fine_tuned_model, tokenizer):
#     count = 0
#     progress = tqdm(total=len(prompts), position=0, leave=True)
#     for prompt in prompts:
#         tokens = tokenizer.apply_chat_template(prompt['prompt'], tokenize=False, add_generation_prompt=True )
#         print(tokens)
#         model_inputs = tokenizer([tokens], return_tensors='pt').to("cuda")
#         generated_ids = fine_tuned_model.generate(**model_inputs, 
#                                        pad_token_id=tokenizer.eos_token_id, 
#                                        max_new_tokens=50,
#                                       temperature=0.001)
#         parsed_ids = [
#             output_ids[len(input_ids):] for input_ids, output_ids in zip(model_inputs.input_ids, generated_ids)
#         ]
#         resp = {
#             'date': prompt['date'],
#             'security': prompt['security'],
#             'response': tokenizer.batch_decode(parsed_ids, skip_special_tokens=True)[0]
#         }
#         outputs.append(resp)
#         progress.update()

# # Load the base model 
# model_helper = mh.ModelHelper('tmp/fs')
# base_model = ir.load_model_from_storage(run_config['model_s3_loc'])
# # clear the local folder once completed loading into memory
# model_helper.clear_folder(run_config['model_s3_loc'])

# # Update the base model with the fine-tuned modules
# fine_tuned_model = PeftModel.from_pretrained(base_model, 'fine_tuned2')
# # Create the tokenizer
# tokenizer = AutoTokenizer.from_pretrained(run_config['model_hf_id'])
# run_finetuned_backtest(prompt_set[:2], fine_tuned_model, tokenizer)

# # save the output
# with open(f"Results/Earnings/results - Qwen3B Finetuned - EPS only.json", 'w') as f:
#     json.dump(outputs, f)
