==> 02B - Frontier Strategy.ipynb <==
# Frontier Models

This section uses AWS Bedrock to run inference tests using the Claude and Open AI models. We use LangChain and Python multi-threading to run the inference tasks and quickly loop through all of the prompts to generate our output.

import json
import boto3
import botocore
import datetime
import importlib
import prompts
from tqdm import tqdm

from prompts import SYSTEM_PROMPTS
from IPython.display import Markdown, display
import constructors.langchain_strategy as ls


importlib.reload(prompts)
importlib.reload(ls)

# boto3 must be version 1.34.51 or higher
boto3.__version__

# Bedrock client initialise
config = botocore.config.Config(read_timeout=1000)
boto3_bedrock = boto3.client('bedrock-runtime',config=config)

## Strategy Construction for Frontier Models

frontier = ls.FrontierRun(run_name='Claude 3.7',
                       model_id='us.anthropic.claude-3-7-sonnet-20250219-v1:0',
                       dataset_id='dow_quarterly_ltm_v3.json',system_prompt=prompts.SYSTEM_PROMPTS['BASE_CLAUDE']['prompt'],boto3_bedrock=boto3_bedrock)

prompts = frontier.setup_prompts()

frontier.run()

## Initial Bedrock Test
This is an initial run of Bedrock with Anthropic to see how it responds to the financial analysis task as a single model run. The below code was used for the development of the strategy construction module for LangChain


# set the model
model = 'us.anthropic.claude-3-7-sonnet-20250219-v1:0'
# load prompts into memory
with open('Data/prompts.json', 'rb') as f:
    prompts = json.load(f)

def setup_claude_request(prompt: str) -> str:
    
    native_request = {
        "anthropic_version": "bedrock-2023-05-31",
        "max_tokens": 5000,
        "temperature": 0.7,
        "messages": [
            {
                "role":"user",
                "content": [{"type":"text", "text":prompt}]
            }
        ]
    }
    
    return json.dumps(native_request)

GUARDRAIL_ID = os.environ['BEDROCK_GUARDRAIL_ID']
GUARDRAIL_VERSION = os.environ['BEDROCK_GUARDRAIL_VERSION']

model_id = model


def invoke_model_claude(prompt: dict, model_id: str) -> str:
    """Invoke the model using Bedrock. This is specifically designed for Anthropic models"""
    accept = 'application/json'
    contentType = 'application/json'

    # set up the request
    request = setup_claude_request(prompt)
    try:
        
        response = boto3_bedrock.invoke_model(
            body=request, 
            modelId=model_id, 
            accept=accept, 
            contentType=contentType,
            trace = "ENABLED"
        )
        response_body = json.loads(response.get('body').read())
        return response_body.get('content')[0].get('text') #Anthropic response template
    
    except botocore.exceptions.ClientError as error:
        
        if error.response['Error']['Code'] == 'AccessDeniedException':
               print(f"\x1b[41m{error.response['Error']['Message']}\
                    \nTo troubeshoot this issue please refer to the following resources.\
                     \nhttps://docs.aws.amazon.com/IAM/latest/UserGuide/troubleshoot_access-denied.html\
                     \nhttps://docs.aws.amazon.com/bedrock/latest/userguide/security-iam.html\x1b[0m\n")
            
        else:
            raise error

response = invoke_model_claude(prompt, model_id)

display(Markdown(response))

## LangChain version
This is the langgraph version in prep for multi-agentic AI system

### Run Inference Loop for Dow Jones


from langchain.prompts import PromptTemplate
from langchain_core.runnables import RunnableSequence
from langchain_core.rate_limiters import InMemoryRateLimiter

from langchain_aws import ChatBedrock
from pydantic import BaseModel, Field

import concurrent.futures

# set up the LLM in Bedrock
rate_limiter = InMemoryRateLimiter(
    requests_per_second=50,
    check_every_n_seconds=1,
    max_bucket_size=10,
)

llm = ChatBedrock(
    client = boto3_bedrock,
    #model = "us.meta.llama3-1-70b-instruct-v1:0",
    model = 'us.anthropic.claude-3-7-sonnet-20250219-v1:0',
    temperature = 0.5,#0.7,
    max_tokens=4000,
    rate_limiter = rate_limiter
)

system_prompt = SYSTEM_PROMPTS['COT_EARN_CLAUDE']['prompt']
prompt_template = PromptTemplate.from_template(system_prompt)

def run_model(prompt: dict, llm: RunnableSequence) -> dict:
    prompt_in = prompt_template.format(financials=prompt['prompt'][1]['content'])
    output = llm.invoke(prompt_in)
    decision_dict = {
        'date': prompt['date'],
        'security': prompt['security'],
        'response': output
    }
    return decision_dict

#prompt_template.format(financials=prompts[0]['prompt'][1]['content'])
prompt_template

run_model(prompts[0], llm)

%%time
# This creates the initial response - run this in parallel with 100 executors
data = []
with concurrent.futures.ThreadPoolExecutor(max_workers=100) as executor:
    futures = [executor.submit(run_model, prompt, llm) for prompt in prompts]
    data = [f.result() for f in futures]

len(data)

cleaned_output = [{'date': i['date'], 
             'security': i['security'], 
             'response': i['response'].content} 
            for i in data]

with open('Data/Results - Bedrock2.json', 'w') as f:
    json.dump(cleaned_output, f)

## Convert to JSON format
This is needed to run the trade analytics and compare to other LLMs

rate_limiter = InMemoryRateLimiter(
    requests_per_second=10,
    check_every_n_seconds=1,
    max_bucket_size=10,
)

llm = ChatBedrock(
    client = boto3_bedrock,
    #model = "us.meta.llama3-1-70b-instruct-v1:0",
    model = 'us.anthropic.claude-3-7-sonnet-20250219-v1:0',
    temperature = 0.5,#0.7,
    max_tokens=4000,
    rate_limiter = rate_limiter
)

# STEP 1: Set up the structured output
# class FinancialOutput(BaseModel):
#     """Answer to the users question along with justification"""
#     decision: str = Field(..., description="Final Decision: BUY, SELL or HOLD")
#     reason: str = Field(..., description="Summary of the final decision")
#     confidence: int = Field(..., description="How confident you are of the decision")

# with open('Results/Earnings/results - deepseek - cot - 2025-03-30 .828629.json', 'rb') as f:
#     raw_output = json.load(f)

class FinancialOutput(BaseModel):
    """Answer to the users question along with justification"""
    direction: str = Field(..., description="earnings will increase or decrease or stay flat")
    magnitude: str = Field(..., description="size of the increase or decrease")
    reason: str = Field(..., description="Summary of the final decision")
    confidence: str = Field(..., description="How confident you are of the decision")

cleaned_output = raw_output['results']

# STEP 2: set up the system prompt and set the structured output format
system_prompt = "You are an assistant to a financial analyst. You are responsible for formatting the documents that the analyst produces into a machine readable format. Use only the information provided in the context. Convert it into the structured output. Do not add anything to the analysts report and do not change the recommendation. Do not hallucinate. Find the investment decision. Find the conclusion. Add all of the wording of the thought process into the steps section. context: {context}"

prompt_template = PromptTemplate.from_template(system_prompt)
structured_llm = llm.with_structured_output(FinancialOutput)


# STEP 3: Set up the function to call in each thread
# def run_cleanup(prompt: dict, llm: RunnableSequence) -> dict:
#     thought_process = prompt['response']
#     prompt_in = prompt_template.format(context=thought_process)
#     response = llm.invoke(prompt_in)
#     final_dict = {
#         'date': prompt['date'],
#         'security': prompt['security'],
#         'response': {
#             'decision': response.decision,
#             'reason': response.reason,
#             'confidence': response.confidence,
#             'thought_process': thought_process
#         }
#     }
#     return final_dict

def run_cleanup(prompt: dict, llm: RunnableSequence) -> dict:
    thought_process = prompt['response']
    prompt_in = prompt_template.format(context=thought_process)
    response = llm.invoke(prompt_in)
    final_dict = {
        'date': prompt['date'],
        'security': prompt['security'],
        'response': {
            'analyst_direction': response.direction,
            'analyst_magnitude': response.magnitude,
            'reason': response.reason,
            'confidence': response.confidence,
            'thought_process': thought_process
        }
    }
    return final_dict

# STEP 4: 

responses = []
with concurrent.futures.ThreadPoolExecutor(max_workers=10) as executor:
    futures = [executor.submit(run_cleanup, prompt, structured_llm) for prompt in cleaned_output]
    responses = [f.result() for f in futures]

# # STEP 5: convert to the correct format
final_json = {
    'run_date': str(datetime.datetime.now()),
    'system_prompt': SYSTEM_PROMPTS['COT_EARN_CLAUDE']['prompt'],
    'dataset': 'data_quarterly_pit_indu_blended',
    'model': model,
    'results': responses
}

# final_json = {
#     'run_date': str(datetime.datetime.now()),
#     'system_prompt': raw_output['system_prompt'],
#     'dataset': raw_output['dataset'],
#     'model': raw_output['model'],
#     'results': responses
# }

# STEP 6: Save the results 
with open(f'Results/results - Deepseek - COT C -{str(datetime.datetime.now())}.json', 'w') as f:
    json.dump(final_json, f)
