==> 03A - Finetuning.ipynb <==
# Fine-tuning Open Source Model
Fine-tune the Qwen 2.5 3B model on 8,000 prompts from the S&P that do not appear in the Dow Jones Index

%package install trl 

import torch
from transformers import AutoModelForCausalLM, AutoTokenizer, TrainingArguments, BitsAndBytesConfig
from peft import LoraConfig
from trl import SFTTrainer, SFTConfig
from tqdm import tqdm
from datasets import load_dataset
from utils.s3_helper import S3Helper
import random
import bql
import json

from company_data import SecurityData

import importlib
import company_data
import utils.model_helper as mh

importlib.reload(company_data)
bq = bql.Service()

### Set up training data

Need to be careful to remove the securities that will be used to test the strategy to make sure there is no contamination of the datasets. The Training data will use the S&P 500 index as the universe with the Dow Jones securities removed from the universe.

# get the actual Dow Jones datasets
dow_data = company_data.SecurityData('tmp/fs','data_quarterly_pit_indu_refresh_blended.json')
dow_secs = dow_data.get_unique_securities()

# get the training data from S&P
spx_data = company_data.SecurityData('tmp/fs','data_quarterly_pit_spx_refresh_blended.json') #'data_annual_pit_spx.json')#
spx_secs = spx_data.get_unique_securities()

#STEP 2: Get the unique securities that could be part of the training data set
training_secs = []
for sec in spx_secs:
    if sec not in dow_secs:
        training_secs.append(sec)

# STEP 3: select random securities and dates for the training set
dates = spx_data.get_dates()

random.seed(21)
random.shuffle(dates)

training_system_prompt = """You are a financial analyst. Use the following income statement, balance sheet to estimate the Basic EPS for the next fiscal period. Use only the data in the prompt. Provide a confidence score for how confident you are of the decision. If you are not confident then lower the confidence score."""

# assemble the prompts - 1000 dates
prompts = []
count = 10000
for date in dates:
    # Pull out the securities reporting on that date
    securities = spx_data.get_securities_reporting_on_date(date)
    # Loop through the securities
    for security in securities:
        # check to see if the security is in the training set
        if security in training_secs:
            if len(prompts) < count:
                prompt = spx_data.get_prompt(date, security, training_system_prompt)
                #record = {'security': security, 'date': date, 'prompt': training_system_prompt + prompt[1]['content'] + "\nAnswer in JSON format with the next period EPS, the direction, the magnitude and a confidence."}
                record = {'security': security, 'date': date, 'prompt': training_system_prompt + prompt[1]['content'] + "\nThe next period EPS is "}
                prompts.append(record)
            else:
                break

prompts[0]

# Get data from BQL for the actual and estimated EPS for each period
def get_eps_values(security:str, as_of_date:str) -> tuple[float, float]:
    """Return the actual and estimate EPS for a security and point in time date"""
    field = {'eps': bq.data.is_basic_eps_cont_ops(dates=as_of_date, 
                               fpo=bq.func.range('0Q','+1Q'), 
                               currency='USD', 
                               fpt='LTM', 
                               fa_period_year_end='C1231',
                               fa_period_type_source='Q', 
                               fa_act_est_data='AE')}
    # Request the data and unpack to a DataFrame
    req = bql.Request(security,field)
    dt = bq.execute(req)
    df = dt[0].df()
    # extract the actual and estimate
    actual_value = df['eps'][0]
    estimate_value = df['eps'][1]
    return actual_value, estimate_value

# Function to create a JSON response. Not needed for EPS measure
def construct_prompt_response(actual_value, est_value):
    earnings = 'INCREASING' if est_value > actual_value else 'DECREASING'
    magnitude = est_value / actual_value - 1
    if abs(magnitude) < 0.02 :
        mag_descr = 'FLAT'
    elif abs(magnitude) <= 0.05:
        mag_descr = 'SMALL'
    elif abs (magnitude) <= 0.1:
        mag_descr = 'LARGE'
    else:
        mag_descr = 'VERY LARGE'
    return json.dumps({'EPS': round(est_value,3), 'earnings': earnings, 'magnitude':mag_descr})



construct_prompt_response(3,2)

for prompt in prompts:
    actual_value, est_value = get_eps_values(prompt['security'],prompt['date'])
    #
    prompt['eps'] = str(round(est_value,3))
    #prompt['eps'] = construct_prompt_response(actual_value, est_value)

prompts[0]

with open('/tmp/training_prompts4.json', 'w') as f:
    json.dump(prompts, f)

s3_helper = S3Helper('tmp/fs')

#s3_helper.add_file(local_filename='/tmp/training_prompts4.json')
#s3_helper.list_folder()

s3_helper.get_file(filename='training_prompts4.json', local_filename='/tmp/training_prompts4.json')

### Load the LLM for fine tuning

with open('/tmp/training_prompts4.json', 'rb') as f:
    prompts = json.load(f)

# convert to training sets needed jsonl format
data_set_for_training = []
for prompt in prompts:
    #data_set_for_training.append({'prompt':prompt['prompt'][:prompt['prompt'].find('Balance Sheet:')] + "Answer in JSON format with the next period EPS, the direction, the magnitude and a confidence:", 'completion': str(prompt['eps'])})
    data_set_for_training.append({'prompt':prompt['prompt'][:prompt['prompt'].find('Balance Sheet:')] + "\nThe next period EPS is:", 'completion': str(prompt['eps'])})

with open('/tmp/training_set.jsonl', 'w') as f:
     json.dump({'data': data_set_for_training}, f)


dataset = load_dataset("json", data_files="/tmp/training_set.jsonl", field='data', split='train').train_test_split(test_size=0.2)

train_dataset = dataset['train']
test_dataset = dataset['test']

model_loader = mh.ModelHelper('tmp/fs')
model = model_loader.load_model('qwen3b','auto')

model_loader.clear_folder('qwen3b')

tokenizer = AutoTokenizer.from_pretrained('Qwen/Qwen2.5-3B-Instruct', trust_remote_code=True)
tokenizer.pad_token = tokenizer.eos_token
tokenizer.padding_side = 'right'

model.generation_config.pad_token_id = tokenizer.pad_token_id

## Set up for training

# mask the majority of the prompt for estimating
from trl import DataCollatorForCompletionOnlyLM
template = "The next period EPS is:"
#collator = DataCollatorForCompletionOnlyLM(tokenizer.encode(template, add_special_tokens = False)[2:], tokenizer=tokenizer)
collator = DataCollatorForCompletionOnlyLM(template, tokenizer=tokenizer)

lora_parameters = LoraConfig(
    lora_alpha = 8,
    lora_dropout = 0.1,
    target_modules = ['q_proj','v_proj','k_proj', 'o_proj'],
    r = 4,
    bias = 'none',
    task_type = 'CAUSAL_LM'
)

# lora_parameters = LoraConfig(
#     lora_alpha = 16,
#     lora_dropout = 0.1,
#     target_modules = ['q_proj','v_proj','k_proj', 'o_proj'],
#     r = 8,
#     bias = 'none',
#     task_type = 'CAUSAL_LM'
# )

train_parameters = SFTConfig(
    output_dir = '/tmp/qwen_trained',
    num_train_epochs=1,
    per_device_train_batch_size=1,
    per_device_eval_batch_size=1,
    eval_strategy='no',
    gradient_accumulation_steps=1,
    optim='paged_adamw_32bit',
    learning_rate= 1e-4,
    lr_scheduler_type = 'cosine',
    warmup_ratio = 0.03,
    save_steps=500,
    save_total_limit=10,
    logging_steps=5,
    weight_decay=0.001,
    max_grad_norm=0.3,
    max_steps=-1,
    push_to_hub=False,
    save_strategy="steps",
    max_seq_length=4000,
    dataset_text_field='prompt',
    group_by_length=True,
    run_name='Earnings',
    #label_names=['labels']
)

fine_tuning = SFTTrainer(
    model=model,
    train_dataset=train_dataset,
    peft_config=lora_parameters,
    tokenizer=tokenizer,
    args=train_parameters,
    data_collator=collator
)

train_dataset[0]

### Start training the model

fine_tuning.train()

model.get_memory_footprint() / 1e9

# save model to s3
fine_tuning.model.save_pretrained('fine_tuned_json')
