==> 03B - Agentic Strategy Construction.ipynb <==
# Agentic Strategy Construction Tool
This is to produce trade reports that can be used in the evaluation section to backtest. An output example is at the bottom of this notebook to show the agentic decision making process in more detail.

import copy
import pandas as pd
import requesters.company_data as cd
import agents.financial_agent as fa
import agents.committee_agent as ca
import importlib
from tqdm import tqdm
import json

from utils.s3_helper import S3Helper
from datetime import datetime

import concurrent.futures


importlib.reload(fa)
importlib.reload(ca)
importlib.reload(cd)

### Collect the data needed to run the Agents
The Agents will require financial statement datasets and news datasets in order to conduct their analysis.

# Get the company reports
security_datasets = cd.SecurityData('tmp/fs', 'dow_quarterly_ltm_v3.json')
# get the saved news datasets
s3_helper = S3Helper('tmp/fs')
s3_helper.get_file(filename='dow_headlines.parquet', local_filename='/tmp/dow_headlines.parquet')
# Convert to pandas dataframe
news_headlines = pd.read_parquet('/tmp/dow_headlines.parquet')

### Setup the Agentic Models
Running the inference tasks across 893 date/ security combinations in 1:32h.

financial_agent = fa.FinancialAnalystAgent()
committee_agent = ca.CommitteeAgent()

def run_single(security: str, as_of_date: str) -> dict:
    """
    Function to run a single run of the Agent
    """
    company_data = security_datasets.get_security_all_data(as_of_date, security)
    # Time the run
    start_time = datetime.now()
    # Run the financial analyst agent
    financial_report = financial_agent.run(security_data=company_data, 
                                       news_data=news_headlines, 
                                       as_of_date=as_of_date)
    # Run the committee agent
    committee_report = committee_agent.run(senior_analyst_report=financial_report['senior_report'],
                                           financial_statement_analysis=financial_report['financial_report'],
                                           security_data=company_data)
    end_time = datetime.now()
    decision_dict = {
        'date': as_of_date,
        'security': security,
        'earning_decision': financial_report['final_output'].direction,
        'earning_magnitude': financial_report['final_output'].magnitude,
        'earning_confidence': financial_report['final_output'].confidence,
        'recommendation': committee_report['results'].recommendation,
        
        'responses': {'financial_analyst': financial_report,
                     'committee_report': committee_report},
        'time': str(end_time - start_time)
    }
    return decision_dict
    
    

# Set up the securities and security information 
dates_and_securities = security_datasets.date_security_timeseries()


### Run a single security

%%time
# Pick a test security
test_security = dates_and_securities[20]
# Run for a single date/ security
test_output = run_single(test_security['security'],test_security['date'])

#dict_keys(['date', 'security', 'earning_decision', 'earning_magnitude', 'earning_confidence', 'recommendation', 'responses', 'time'])
test_output['recommendation']

### Run the backtest
This will run across multiple threads to speed up the process. Results are collected in the data_output list.

data_output = []
def backtest(dates_and_securities):
    # create the securities
    progress = tqdm(total=len(dates_and_securities), position=0, leave=True)
    # Run on 20 threads to speed up execution
    with concurrent.futures.ThreadPoolExecutor(max_workers=20) as executor:
        futures = [executor.submit(run_single, single['security'], single['date']) for single in dates_and_securities]
        for f in concurrent.futures.as_completed(futures):
            progress.update(n=1)
            data_output.append(f.result())
        

# Run the backtest
backtest(dates_and_securities)

len(data_output)

### Save the raw output to S3

data_analysis_final = copy.deepcopy(data_output)

# Some of the items are not serializable. Convert to string before JSON
for item in data_analysis_final:
    item['responses']['financial_analyst']['final_output'] = str(item['responses']['financial_analyst']['final_output'])
    item['responses']['committee_report']['results'] = str(item['responses']['committee_report']['results'])
    item['responses']['committee_report']['history'] = str(item['responses']['committee_report']['history'])
    item['time']=str(item['time'])

with open('/tmp/agentic_output_llama_v6.json', 'w') as f:
    json.dump(data_analysis_final, f)

s3_helper.add_file(local_filename='/tmp/agentic_output_llama_v6.json')

### Create the Trade Report

def trade_report_generation(data, trade_only=False) -> pd.DataFrame:
    """
    Function to create the trade report to pass into the Strategy Analysis tool
    """
    trade_report = []
    if trade_only:
        for item in data:
            trade_report.append({'date': item['date'], 
                                 'security': item['security'], 
                                 'decision': item['recommendation'],
                                 'confidence': item['earning_confidence']})
    
        return pd.DataFrame(data=trade_report)
    else:
        for item in data:
            trade_report.append({'date': item['date'], 
                                 'security': item['security'], 
                                 'decision': item['recommendation'],
                                 'confidence': item['earning_confidence'],
                                 'earning_decision': item['earning_decision'],
                                 'earning_magnitude': item['earning_magnitude']})
        return trade_report

trades = trade_report_generation(data_analysis_final)

with open('Results/Agentic/trades_llama_v7.json', 'w') as f:
    json.dump(trades, f)

### Example Output from agent

import json
from IPython.display import Markdown, display
from utils.s3_helper import S3Helper

s3_helper = S3Helper('tmp/fs')
s3_helper.get_file('agentic_output_llama_v6.json','/tmp/agentic_output_llama_v6.json')

with open('/tmp/agentic_output_llama_v6.json', 'rb') as f:
    all_data = json.load(f)

security = all_data[10]
response = security['responses']

security['date']

### Financial Report

display(Markdown(response['financial_analyst']['financial_report']))

## Financial Agent
### News Summary

# Headlines
response['financial_analyst']['cleaned_headlines']

display(Markdown(response['financial_analyst']['news_report']))

### Senior Analyst Report

display(Markdown(response['financial_analyst']['senior_report']))

## Committee Agent
### Buy Summary

display(Markdown(response['committee_report']['summary_buy']))

### Sell Summary

display(Markdown(response['committee_report']['summary_sell']))

### Hold Summary

display(Markdown(response['committee_report']['summary_hold']))

### Debate

response['committee_report']['history']

### Final result from the Committee Agent

response['committee_report']['results']

### Unused code

# from utils.s3_helper import S3Helper

# s3_helper = S3Helper('tmp/fs')
# s3_helper.get_file(filename='agentic_output_llama_v3_1.json',local_filename='/tmp/agentic_output_llama_v3_1.json')

# import json
# with open('/tmp/agentic_output_llama_v3_1.json', 'rb') as f:
#     data_output = json.load(f)

# print(data_output[1])


s3_helper.get_file('agentic_output_llama_v6.json','/tmp/agentic_output_llama_v6.json')
