==> 05 - Agentic Tests.ipynb <==
# Multi-Agentic Financial Analyst
This notebook will be the test build of the multi-agentic AI system. This is a notebook to experiment with the solution and to 

#%package install boto3=1.34.51 langchain 

import bql
import boto3
import botocore
import random
import json
import importlib
from utils.s3_helper import S3Helper

from langchain.prompts import PromptTemplate
from langchain_core.runnables import RunnableSequence
from langchain_core.rate_limiters import InMemoryRateLimiter

from langchain_aws import ChatBedrock
from pydantic import BaseModel, Field
from IPython.display import Markdown, display
import company_data
from datetime import datetime
from dateutil.relativedelta import relativedelta

import prompts
import pandas as pd

import concurrent.futures

bq = bql.Service()

importlib.reload(prompts)

# Bedrock client initialise
config = botocore.config.Config(read_timeout=1000)
boto3_bedrock = boto3.client('bedrock-runtime',config=config)

with open('Data/prompts.json', 'rb') as f:
    prompt_set = json.load(f)

### Configure the model

model_claude_id = 'us.anthropic.claude-3-7-sonnet-20250219-v1:0'
model_claude_small = "us.anthropic.claude-3-5-haiku-20241022-v1:0"
model_llama_id = 'us.meta.llama3-1-70b-instruct-v1:0'
model_llama_small_id = "us.meta.llama3-1-8b-instruct-v1:0"

# set up the LLM in Bedrock
rate_limiter = InMemoryRateLimiter(
    requests_per_second=50,
    check_every_n_seconds=1,
    max_bucket_size=10,
)

llm = ChatBedrock(
    client = boto3_bedrock,
    model_id = model_claude_id,
    temperature = 0.01,
    max_tokens=4000,
    rate_limiter = rate_limiter
)

llm_small = ChatBedrock(
    client = boto3_bedrock,
    model_id = model_llama_small_id,
    temperature = 0.01,
    max_tokens = 4000,
    rate_limiter = rate_limiter
)

llm_debate = ChatBedrock(
    client = boto3_bedrock,
    model_id = model_claude_small,
    temperature = 0.01,
    max_tokens = 4000,
    rate_limiter = rate_limiter
)

# test prompt with a single security
security = prompt_set[0]
sec_name = security['security']
sec_date = security['date']
sec_fs = security['prompt'][1]['content']
sec_prices = sec_fs[sec_fs.find('Historical Price:') + 17:]

system_prompt = prompts.SYSTEM_PROMPTS['COT_CLAUDE_TEST']['prompt']
prompt_template = PromptTemplate.from_template(system_prompt)
# Create the prompt to feed into the model
prompt_in = prompt_template.format(financials=sec_fs)

output_financial_statements = llm.invoke(prompt_in)


output_financial_statements.content
#display(Markdown(output.content))

# Analyse the News datasets

s3_helper = S3Helper('tmp/fs')
s3_helper.get_file(filename='dow_headlines.parquet', local_filename='/tmp/dow_headlines.parquet')
news_headlines = pd.read_parquet('/tmp/dow_headlines.parquet')

news_headlines

# Convert sec_name to figi
def convert_to_figi(sec_name: str) -> str:
    """Function to convert Bloomberg tickers in a dataframe to FIGIs for ESL"""
    univ      = sec_name
    field     = {'figi': bq.data.composite_id_bb_global(), 
                 'name': bq.data.name(), 
                 'sector': bq.data.bics_level_1_sector_name()}
    req = bql.Request(univ, field)
    data = bq.execute(req)
    return (data[0].df().loc[sec_name]['figi'], data[1].df().loc[sec_name]['name'], data[2].df().loc[sec_name]['sector'])
    

figi_name, name, sector = convert_to_figi(sec_name)
figi_name

sector


def filter_by_company_date(dataset, security, max_date = None):
    
    if max_date != None:
        #claculate the minimum date to get a 3 month window
        min_date = datetime.strptime(max_date,"%Y-%m-%d") + relativedelta(months=-3)
        min_date = min_date.strftime("%Y-%m-%d")
        filtered_dataset = dataset[(dataset['TimeOfArrival'] < max_date) &  (dataset['TimeOfArrival'] >= min_date) & (dataset['Assigned_ID_BB_GLOBAL'] == security)]
    else:
        filtered_dataset = dataset[dataset['Assigned_ID_BB_GLOBAL'] == security]
    
    filtered_list = filtered_dataset['Headline'].to_list()

    if len(filtered_list) >= 50:
        return random.sample(filtered_list, 50)
    else:
        return filtered_list
    

sec_date

headline_filter_test = filter_by_company_date(news_headlines, figi_name, sec_date)

headline_filter_test

system_prompt = """You are an assistant to a financial analyst analyzing {security} You must remove any reference to {security} and their products from the following list of headlines and replace them with the term 'blah'. Replace the names of any people such as ceo in the article with the term 'whah' do not refer to {security} at all in your answer:{headlines}"""
prompt_template = PromptTemplate.from_template(system_prompt)
# Create the prompt to feed into the model
prompt_in = prompt_template.format(headlines=headline_filter_test, security=name)

prompt_in

output_cleaned = llm_small.invoke(prompt_in)

output_cleaned

output_ceo = llm.invoke("Guess the name of the company blah from the following headline: blah ceo whah says company had slow start to 2020")

output_ceo

cleaned_headlines = output_cleaned.content.split('\n')

cleaned_headlines[1:]

system_prompt = """You are a financial analyst and are reviewing news for company called blah over the last three months. Blah is in the {sector} sector. Start by listing the revenue drivers for the sector. Then look through the below headlines and determine if blah will see an increase or decrease in their earnings over the next quarter. Think through your response. {headlines}"""
prompt_template = PromptTemplate.from_template(system_prompt)
# Create the prompt to feed into the model
prompt_in = prompt_template.format(headlines=cleaned_headlines[1:], sector=sector)

prompt_in

output_news = llm.invoke(prompt_in)

display(Markdown(output_news.content))

system_prompt = """You are a senior financial analyst and review your teams work. You are looking at a financial summary and news for 'blah'. Using the summaries only, critique the report and construct an alternative narrative. If the narrative is in agreement with the two reports, make clear your belief in the direction of earning. If in disagreement, state why you disagree. Think through your response. {financial_summary} \n {news_summary}"""
prompt_template = PromptTemplate.from_template(system_prompt)
# Create the prompt to feed into the model
prompt_in = prompt_template.format(financial_summary=output_financial_statements.content, news_summary=output_news.content)

prompt_in

final_output = llm.invoke(prompt_in)

display(Markdown(final_output.content))

## Langgraph Debate Step

from langgraph.graph import START, END, StateGraph
from langgraph.graph.message import add_messages

from typing import Dict, TypedDict, Optional

def llm_debate_invoke (prompt):
    return llm_debate.invoke(prompt).content

# number of agents
agents = ['BUY', 'SELL', 'HOLD']

class GraphState(TypedDict):
    classification: Optional[str] = None
    last_agent: Optional[str] = None
    history: Optional[str] = None
    summary_buy: Optional[str] = None
    summary_sell: Optional[str] = None
    summary_hold: Optional[str] = None
    current_response: Optional[str] = None
    count: Optional[int] = None
    results: Optional[str] = None
    consensus: Optional[dict[str,str]]

analyst_workflow = StateGraph(GraphState)

debate_agent_initial_system_prompt = """You are part of the investment committee at an asset management firm. Your senior analyst in the {sector} sector has presented the findings of reports with your fellow committee members until you come to an agreement on whether to recommend BUY, SELL or HOLD based on the stock price below and the reports from the analysts. If you are all in agreement then stop. You are advocating to {decision} the stock. Give a short 200 word summary of your decision. Only use the following context. Think through the response. Context: {senior_analyst_report} {financial_statement} {stock_prices}"""
agentic_initial_template = PromptTemplate.from_template(debate_agent_initial_system_prompt)

debate_agent_system_prompt = """You are part of the investment committee at an asset management firm. Your senior analyst in the {sector} sector has presented the findings of reports with your fellow committee members until you come to an agreement on whether to recommend {decision}. Do not repeat any previous arguments. Base your response only on the report and current conversation. If you do not recommend {decision}, state which direction you advocate. Provide the argument in less than 200 words. Think through your analysis. For review: {senior_analyst_report} \nPrior conversation: {conversation} """
agentic_debate_template = PromptTemplate.from_template(debate_agent_system_prompt)

debate_direction_system_prompt = """Based on the response, return the sentiment of the response as BUY, SELL or HOLD. Only return a single word: BUY, SELL or HOLD. Conversation: {conversation}"""
debate_direction_template = PromptTemplate.from_template(debate_direction_system_prompt)

result_system_prompt = """You are part of the investment committee at an asset management firm. You hold the deciding vote on whether to BUY, SELL or HOLD a stock. You must always proceed with the majority decision of the analysts. If there is no majority decision, decide on a winner. There can only be one decision. Only output the decision. The conversation is: {conversation}"""
results_template = PromptTemplate.from_template(result_system_prompt)

def debate_format(state, state_update, decision):
    """
    Function to run the debate between three agents. All use the same template/ format
    state_update: str - pass name of object to store the agent summary
    decision: str - identity of the agent
    """
    summary = state.get('history', '').strip()
    summary_x = state.get(state_update, '').strip()
    current_response = state.get('current_response', '').strip()
    if summary_x=='Nothing':
        # this is the initial argument
        prompt_in = agentic_initial_template.format(decision=decision, 
                                                    senior_analyst_report=final_output.content,
                                                    financial_statement=output_financial_statements.content,
                                                    stock_prices=sec_prices,
                                                    sector=sector)
        argument = decision + ":" + llm_debate_invoke(prompt_in)
        if summary == 'Nothing':
            return {'history': 'START\n' + argument, 
                    state_update: argument, 
                    'current_response': argument, 
                    'count':state.get('count')+1,
                    'last_agent': decision}
        else:
            return {'history': summary + '\n' + argument, 
                    state_update: argument, 
                    'current_response': argument, 
                    'count':state.get('count')+1,
                    'last_agent': decision}
    else:
        # this runs during the debate
        prompt_in = agentic_debate_template.format(decision=decision,
                                                      sector=sector,
                                                      senior_analyst_report=final_output.content,
                                                      conversation=summary)
        argument = decision + ":" + llm_debate_invoke(prompt_in)

        return {'history': summary + '\n' + argument,
                'current_response': argument, 
                'count': state.get('count') + 1,
                'last_agent': decision}

def next_node(state):
    last_agent = state.get('last_agent').strip()
    response_from_agent = state.get('current_response')
    current_consensus = state.get('consensus')
    prompt_in = debate_direction_template.format(conversation=response_from_agent)
    last_node = llm_debate_invoke(prompt_in)
    print(last_agent + ":" + last_node)
    current_consensus[last_agent] = last_node
    return {'consensus': current_consensus}

def handle_buy(state):
    return debate_format(state, 'summary_buy', 'BUY')

def handle_sell(state):
    return debate_format(state, 'summary_sell', 'SELL')

def handle_hold(state):
    return debate_format(state, 'summary_hold', 'HOLD')

def result(state):
    summary = state.get('history').strip()
    prompt_in = results_template.format(conversation=summary)
    return {"results":llm_debate_invoke(prompt_in)}

analyst_workflow.add_node('next_node', next_node)
analyst_workflow.add_node('handle_buy', handle_buy)
analyst_workflow.add_node('handle_sell', handle_sell)
analyst_workflow.add_node('handle_hold', handle_hold)
analyst_workflow.add_node('result', result)

# Add the logic to decide the next steps in the debate
def decide_next_node(state):
    last_agent = state.get('last_agent')
    current_consensus = state.get('consensus')
    if state.get('summary_sell', '') == 'Nothing':
        return 'handle_sell'
    if state.get('summary_hold', '') == 'Nothing':
        return 'handle_hold'

    if len(set(current_consensus.values())) == 1:
        # all the agents are in consensus
        return 'result'
    else:
        # agent is agreeing with their classification.
        # debate should continue
        if last_agent == 'BUY':
            return 'handle_sell'
        if last_agent == 'SELL':
            return 'handle_hold'
        if last_agent == 'HOLD':
            return 'handle_buy'

def check_conv_length(state):
    return "result" if state.get("count")==12 else "next_node"

analyst_workflow.add_conditional_edges(
    "next_node",
    decide_next_node,
    {
        "handle_buy": "handle_buy",
        "handle_sell": "handle_sell",
        "handle_hold": "handle_hold",
        "result": "result"
    }
)

analyst_workflow.add_conditional_edges(
    "handle_buy",
    check_conv_length,
    {
        "result": "result",
        "next_node": "next_node"
    }
)

analyst_workflow.add_conditional_edges(
    "handle_sell",
    check_conv_length,
    {
        "result": "result",
        "next_node": "next_node"
    }
)

analyst_workflow.add_conditional_edges(
    "handle_hold",
    check_conv_length,
    {
        "result": "result",
        "next_node": "next_node"
    }
)

analyst_workflow.set_entry_point("handle_buy")
analyst_workflow.add_edge('result', END)

app = analyst_workflow.compile()

app.get_graph()

%%time
initial_state = {
    'count':0,
    'history':'Nothing',
    'current_response':'',
    'summary_buy': 'Nothing',
    'summary_sell': 'Nothing',
    'summary_hold': 'Nothing',
    'consensus': {'BUY':'BUY', 'SELL':'SELL', 'HOLD':'HOLD'}
}

conversation = app.invoke(initial_state)

print(conversation['results'])

print(conversation['history'])

print(conversation['consensus'])

conversation['results']

univ      = ['IBM US Equity', 'MSFT US Equity']
field     = {'figi': bq.data.composite_id_bb_global(), 
                 'name': bq.data.name(), 
                 'sector': bq.data.bics_level_1_sector_name()}
req = bql.Request(univ, field)
data = bq.execute(req)

df = [d.df() for d in data]
#df1 = pd.merge(on=)

pd.concat(df, axis=1)

c = {'BUY':'BUY', 'SELL':'SELL', 'HOLD':'HOLD'}

c = {'BUY':'BUY', 'SELL':'BUY', 'HOLD':'HOLD'}

len(set(c.values()))

max_date = '2020-03-31'

date = datetime.strptime(max_date,"%Y-%m-%d")
min_date = date + relativedelta(months=-3)
min_date.strftime("%Y-%m-%d")

news_headlines[news_headlines['TimeOfArrival'] < max_date]

min_date.strftime()

index_members = ['BBG000B9XRY4', 'BBG000BBJQV0', 'BBG000BBS2Y0', 'BBG000BCQZS4',
       'BBG000BCSST7', 'BBG000BF0K17', 'BBG000BH4R78', 'BBG000BJ81C1',
       'BBG000BKZB36', 'BBG000BLNNH6', 'BBG000BMHYD1', 'BBG000BMX289',
       'BBG000BN2DC2', 'BBG000BNSZP1', 'BBG000BP52R2', 'BBG000BPD168',
       'BBG000BPH459', 'BBG000BR2B91', 'BBG000BR2TH3', 'BBG000BSXQV7',
       'BBG000BVPV84', 'BBG000BW8S60', 'BBG000BWLMJ4', 'BBG000BWXBC2',
       'BBG000C0G1D1', 'BBG000C3J3C9', 'BBG000C5HS04', 'BBG000C6CFJ5',
       'BBG000CH5208', 'BBG000DMBXR2', 'BBG000GZQ728', 'BBG000H556T9',
       'BBG000HS77T5', 'BBG000K4ND22', 'BBG000PSKYX7', 'BBG00BN96922']



field = bq.data.composite_id_bb_global()
req = bql.Request(index_members, field)
data = bq.execute(req)
data[0].df()
