==> 05 - Langgraph Test.ipynb <==
# LangGraph Analyst Workflow Test
This notebook uses two agents in LangGraph: The Financial Analysis Agent and the Committee debating agent. 

The notebook has two components. The first is an implementation and test of the agent classes. The second is the implementation in the notebook used for design and testing purposes. 

!pip -q install --force-reinstall langgraph langchain_aws langchain-huggingface langchain

## LangGraph Implementation with Objects

import company_data as cd
import agents.financial_agent as fa
import agents.committee_agent as ca
import importlib
from utils.s3_helper import S3Helper
import pandas as pd

importlib.reload(fa)
importlib.reload(ca)

#### Get the security Data

# Request financial Data
sec_data = cd.SecurityData('tmp/fs', 'dow_quarterly_ltm_v2.json')
dates_and_securities = sec_data.date_security_timeseries()
test_security = dates_and_securities[10]
security_data = sec_data.get_security_all_data(test_security['date'], test_security['security'])
#security_data = sec_data.get_security_all_data('2024-08-05', 'JNJ UN Equity')


# Request News data
s3_helper = S3Helper('tmp/fs')
s3_helper.get_file(filename='dow_headlines.parquet', local_filename='/tmp/dow_headlines.parquet')
news_headlines = pd.read_parquet('/tmp/dow_headlines.parquet')

#### Run with the Financial Agent

agent = fa.FinancialAnalystAgent()

%%time
conversation = agent.run(security_data, news_headlines, test_security['date'])

#dict_keys(['company_details', 'financial_report', 'cleaned_headlines', 'news_report', 'senior_report', 'final_output'])
conversation['senior_report']

#### Pass to the Committee Agent

committee = ca.CommitteeAgent()

committee_decision = committee.run(conversation['senior_report'],
                                  conversation['financial_report'],
                                  security_data)

#dict_keys(['last_agent', 'history', 'summary_buy', 'summary_sell', 'summary_hold', 'current_response', 'count', 'results', 'consensus', 'senior_analyst_report', 'financial_statement_analysis', 'sector', 'stock_prices'])
committee_decision['results']

print(committee.get_graph().draw_mermaid())

print(agent.get_graph().draw_mermaid())

## Full notebook implementation
The code below is the notebook implementation of the Financial Agent Class

import bql
import boto3
import botocore
import random
import json
import importlib
from utils.s3_helper import S3Helper

from langchain.prompts import PromptTemplate
from langchain_core.runnables import RunnableSequence
from langchain_core.rate_limiters import InMemoryRateLimiter

from langchain_aws import ChatBedrock
from pydantic import BaseModel, Field
from IPython.display import Markdown, display
import company_data
from datetime import datetime
from dateutil.relativedelta import relativedelta

import prompts
import pandas as pd

import concurrent.futures

from langgraph.graph import START, END, StateGraph
from langgraph.graph.message import add_messages

from typing import Dict, TypedDict, Optional

bq = bql.Service()

# Configure Bedrock
# Bedrock client initialise
config = botocore.config.Config(read_timeout=1000)
boto3_bedrock = boto3.client('bedrock-runtime',config=config)

with open('Data/prompts.json', 'rb') as f:
    prompt_set = json.load(f)

### Configure the Models

model_claude_id = 'us.anthropic.claude-3-7-sonnet-20250219-v1:0'
model_claude_small = "us.anthropic.claude-3-5-haiku-20241022-v1:0"
model_llama_id = 'us.meta.llama3-1-70b-instruct-v1:0'
model_llama_small_id = "us.meta.llama3-1-8b-instruct-v1:0"

# set up the LLM in Bedrock
rate_limiter = InMemoryRateLimiter(
    requests_per_second=50,
    check_every_n_seconds=1,
    max_bucket_size=10,
)

llm_thinker = ChatBedrock(
    client = boto3_bedrock,
    model_id = model_claude_id,
    temperature = 0.01,
    max_tokens=4000,
    rate_limiter = rate_limiter
)

llm_small = ChatBedrock(
    client = boto3_bedrock,
    model_id = model_llama_small_id,
    temperature = 0.01,
    max_tokens = 4000,
    rate_limiter = rate_limiter
)

llm_debate = ChatBedrock(
    client = boto3_bedrock,
    model_id = model_claude_small,
    temperature = 0.01,
    max_tokens = 4000,
    rate_limiter = rate_limiter
)

## Prompts

statement_analysis_system_prompt = prompts.SYSTEM_PROMPTS['COT_CLAUDE_TEST']['prompt']
statement_analysis_template = PromptTemplate.from_template(statement_analysis_system_prompt)

clean_headlines_system_prompt = """You are an assistant to a financial analyst analyzing {security} You must remove any reference to {security} and their products from the following list of headlines and replace them with the term 'blah'. Replace the names of any people such as ceo in the article with the term 'whah' do not refer to {security} at all in your answer:{headlines}"""
clean_headlines_template = PromptTemplate.from_template(clean_headlines_system_prompt)

company_news_system_prompt = """You are a financial analyst and are reviewing news for company called blah over the last three months. Blah is in the {sector} sector. Start by listing the revenue drivers for the sector. Then look through the below headlines and determine if blah will see an increase or decrease in their earnings over the next quarter. Think through your response. {headlines}"""
company_news_template = PromptTemplate.from_template(company_news_system_prompt)

senior_analysis_prompt = """You are a senior financial analyst and review your teams work. You are looking at a financial summary and news for 'blah'. Using the summaries only, critique the report and construct an alternative narrative. If the narrative is in agreement with the two reports, make clear your belief in the direction of earning. If in disagreement, state why you disagree. Think through your response. {financial_summary} \n {news_summary}"""
senior_analyst_template = PromptTemplate.from_template(senior_analysis_prompt)

def analyst_llm(llm, prompt):
    return llm.invoke(prompt).content

class CompanyData(TypedDict):
    name: str
    figi_name: str
    sector: str
    sec_fs: str
    headlines: list[str]
    stock_prices: str

class GraphState(TypedDict):
    company_details: Optional[dict[str,str]]
    initial_analysis: Optional[str]
    cleaned_headlines: Optional[list]
    news_report: Optional[str]
    senior_report: Optional[str]

analyst_workflow = StateGraph(GraphState)

def financial_statement_analysis(state):
    # Create the prompt to feed into the model
    company_details = state.get('company_details')
    sec_fs = company_details['sec_fs']
    prompt_in = statement_analysis_template.format(financials=sec_fs)
    financial_analysis = analyst_llm(llm_thinker, prompt_in)
    return {'initial_analysis': financial_analysis}

def clean_headlines(state):
    company_details = state.get('company_details')
    unclean_headlines = company_details['headlines']
    # Create the prompt to feed into the model
    prompt_in = clean_headlines_template.format(headlines=unclean_headlines, security=name)
    clean_headlines = analyst_llm(llm_small, prompt_in)
    return {'cleaned_headlines': clean_headlines}

def news_summary(state):
    # Create the prompt to feed into the model
    company_details = state.get('company_details')
    clean_headlines = state.get('cleaned_headlines')
    prompt_in = company_news_template.format(headlines=clean_headlines[1:], sector=company_details['sector'])
    news_summarisation = analyst_llm(llm_thinker, prompt_in)
    return {'news_summary': news_summarisation}

def final_report(state):
    company_details = state.get('company_details')
    initial_analysis = state.get('initial_analysis')
    news_summary = state.get('news_summary')
    prompt_in = senior_analyst_template.format(financial_summary=initial_analysis, news_summary=news_summary)
    final_report_output = analyst_llm(llm_thinker, prompt_in)
    return {'senior_report': final_report_output}

analyst_workflow.add_node('financial_statement_analysis', financial_statement_analysis)
analyst_workflow.add_node('clean_headlines', clean_headlines)
analyst_workflow.add_node('news_summary', news_summary)
analyst_workflow.add_node('final_report', final_report)

analyst_workflow.set_entry_point('financial_statement_analysis')
analyst_workflow.add_edge('financial_statement_analysis', 'clean_headlines')
analyst_workflow.add_edge('clean_headlines', 'news_summary')
analyst_workflow.add_edge('news_summary', 'final_report')
analyst_workflow.add_edge('final_report', END)

app = analyst_workflow.compile()

app.get_graph()

## Test on an example firm

# test prompt
security = prompt_set[0]
sec_name = security['security']
sec_date = security['date']
sec_fs = security['prompt'][1]['content']
sec_prices = sec_fs[sec_fs.find('Historical Price:') + 17:]

# get the news headlines
s3_helper = S3Helper('tmp/fs')
s3_helper.get_file(filename='dow_headlines.parquet', local_filename='/tmp/dow_headlines.parquet')
news_headlines = pd.read_parquet('/tmp/dow_headlines.parquet')

# Convert sec_name to figi
def get_additional_data(sec_name: str) -> str:
    """Function to convert Bloomberg tickers in a dataframe to FIGIs for ESL"""
    univ      = sec_name
    field     = {'figi': bq.data.composite_id_bb_global(), 
                 'name': bq.data.name(), 
                 'sector': bq.data.bics_level_1_sector_name()}
    req = bql.Request(univ, field)
    data = bq.execute(req)
    return (data[0].df().loc[sec_name]['figi'], data[1].df().loc[sec_name]['name'], data[2].df().loc[sec_name]['sector'])

figi_name, name, sector = get_additional_data(sec_name)
    

def filter_by_company_date(dataset, security, max_date = None):
    
    if max_date != None:
        #claculate the minimum date to get a 3 month window
        min_date = datetime.strptime(max_date,"%Y-%m-%d") + relativedelta(months=-3)
        min_date = min_date.strftime("%Y-%m-%d")
        filtered_dataset = dataset[(dataset['TimeOfArrival'] < max_date) &  (dataset['TimeOfArrival'] >= min_date) & (dataset['Assigned_ID_BB_GLOBAL'] == security)]
    else:
        filtered_dataset = dataset[dataset['Assigned_ID_BB_GLOBAL'] == security]
    
    filtered_list = filtered_dataset['Headline'].to_list()

    if len(filtered_list) >= 50:
        return random.sample(filtered_list, 50)
    else:
        return filtered_list

headline_filter_test = filter_by_company_date(news_headlines, figi_name, sec_date)

company_details = CompanyData({'name':name,
                              'figi_name': figi_name,
                              'sector': sector,
                              'sec_fs': sec_fs,
                              'headlines': headline_filter_test,
                              'stock_prices': sec_prices})

%%time
conversation = app.invoke({'company_details': company_details})

conversation['senior_report']
